+++ kernel/kernel/exec.h
@@ -7,4 +7,4 @@
 /// @param pathname Name of ELF file to execute
 /// @param argv Command arguments
 /// @return -1 on failure
-int32_t execv(char *pathname, char **argv);
+ssize_t execv(char *pathname, char **argv);
